+++ Source/Artifacts.Tools/Program.cs
@@ -73,22 +73,11 @@ namespace Dolittle.Artifacts.Tools
                 if (boundedContextConfigurationRetrievalResult == BoundedContextConfigurationUtilities.BoundedContextRetrievalResult.NewBoundedContextConfig)
                     boundedContextConfiguration.Topology = new TopologyConfiguration();
                 
-                var types = assemblyLoader
-                    .GetProjectReferencedAssemblies()
-                    .SelectMany(_ => _.ExportedTypes)
-                    .Where(_ =>
-                        _artifactTypes
-                        .Any(at => at.Type.IsAssignableFrom(_)))
-                    .ToArray();
+                var types = GetArtifactsFromAssembly(assemblyLoader);
 
                 ThrowIfArtifactWithNoModuleOrFeature(types);
 
-                var typePaths = types
-                    .Select(_ => 
-                    string.Join(NamespaceSeperator, _.Namespace.Split(NamespaceSeperator).Skip(1)))
-                    .Where(_ => _.Length > 0)
-                    .Distinct()
-                    .ToArray();
+                var typePaths = ExtractTypePaths(types);
                 
                 ThrowIfContainsInvalidTypePath(typePaths, boundedContextConfiguration.UseModules);
                 
@@ -120,13 +109,10 @@ namespace Dolittle.Artifacts.Tools
                     )
                 );
 
-                ThrowIfDuplicateArtifacts(artifactsConfiguration);
-                WarnIfFeatureMissingFromTopology(artifactsConfiguration, boundedContextConfiguration);
-                WarnIfArtifactNoLongerInStructure(artifactsConfiguration, types);
+                artifactsConfiguration.ValidateArtifacts(boundedContextConfiguration, types);
 
                 var hasChanges = newArtifacts > 0;
 
-                
                 if (missingPaths.Any()) _boundedContextConfigurationManager.Save(boundedContextConfiguration); 
                 if (hasChanges) _artifactsConfigurationManager.Save(artifactsConfiguration);
 
@@ -147,6 +133,40 @@ namespace Dolittle.Artifacts.Tools
 
             return 0;
         }
+
+        static void SetupConfigurationManagers()
+        {
+            var container = new ActivatorContainer();
+            var converterProviders = new FixedInstancesOf<ICanProvideConverters>(new []
+            {
+                new Dolittle.Concepts.Serialization.Json.ConverterProvider()
+            });
+
+            var serializer = new Serializer(container, converterProviders);
+            _boundedContextConfigurationManager = new BoundedContextConfigurationManager(serializer);
+            _artifactsConfigurationManager = new ArtifactsConfigurationManager(serializer);
+        }
+        private static string[] ExtractTypePaths(Type[] types)
+        {
+            return types
+                .Select(_ => 
+                string.Join(NamespaceSeperator, _.Namespace.Split(NamespaceSeperator).Skip(1)))
+                .Where(_ => _.Length > 0)
+                .Distinct()
+                .ToArray();
+        }
+
+        static Type[] GetArtifactsFromAssembly(AssemblyLoader assemblyLoader)
+        {
+            return assemblyLoader
+                .GetProjectReferencedAssemblies()
+                .SelectMany(_ => _.ExportedTypes)
+                .Where(_ =>
+                    _artifactTypes
+                    .Any(at => at.Type.IsAssignableFrom(_)))
+                .ToArray();
+        }
+
         static void ThrowIfArtifactWithNoModuleOrFeature(Type[] types)
         {
             bool hasInvalidArtifact = false;
@@ -175,19 +195,6 @@ namespace Dolittle.Artifacts.Tools
                 }
                 if (hasInvalidTypePath) throw new InvalidArtifact();
             }
-        }
-
-        static void SetupConfigurationManagers()
-        {
-            var container = new ActivatorContainer();
-            var converterProviders = new FixedInstancesOf<ICanProvideConverters>(new []
-            {
-                new Dolittle.Concepts.Serialization.Json.ConverterProvider()
-            });
-
-            var serializer = new Serializer(container, converterProviders);
-            _boundedContextConfigurationManager = new BoundedContextConfigurationManager(serializer);
-            _artifactsConfigurationManager = new ArtifactsConfigurationManager(serializer);
         }
 
         static void AddExistingArtifactPaths(BoundedContextConfiguration boundedContextConfiguration, ref List<string> existingArtifactPaths)
@@ -258,67 +265,10 @@ namespace Dolittle.Artifacts.Tools
 
                         targetProperty.SetValue(artifactsByTypeDefinition, newAndExistingArtifacts);
                     }
+                    
                 }
             });
             return newArtifacts;
         }
-
-        static void WarnIfFeatureMissingFromTopology(ArtifactsConfiguration artifactsConfiguration, BoundedContextConfiguration boundedContextConfiguration)
-        {
-            Dictionary<Feature, FeatureName> featureMap = boundedContextConfiguration.RetrieveAllFeatureIds();
-
-            foreach (var artifact in artifactsConfiguration.Artifacts)
-            {
-                if (!featureMap.ContainsKey(artifact.Key))
-                {
-                    ConsoleLogger.LogWarning($"Found artifacts under a Feature that does not exist in the topology. Feature: {artifact.Key}");
-                    Console.WriteLine("Artifacts:");
-                    
-                    var artifactDefinitions = artifactsConfiguration.GetAllArtifactDefinitions();
-                    foreach (var definition in artifactDefinitions)
-                        Console.WriteLine($"\tArtifact: {definition.Artifact.Value} CLR-type: {definition.Type.TypeString} @{definition.Generation.Value}");
-                    
-                }
-            }
-        }
-        static void WarnIfArtifactNoLongerInStructure(ArtifactsConfiguration artifactsConfiguration, IEnumerable<Type> types)
-        {
-            var artifactDefinitions = new List<ArtifactDefinition>();
-            foreach (var artifactDefinition in artifactsConfiguration.GetAllArtifactDefinitions())
-            {
-                if (!types.Contains(artifactDefinition.Type.GetActualType()))
-                    artifactDefinitions.Add(artifactDefinition);
-                
-            }
-            if (artifactDefinitions.Any())
-            {
-                ConsoleLogger.LogWarning("There are artifacts that are not found in the Bounded Context structure anymore. You may have to write migrators for them:");
-                Console.WriteLine("Artifacts:");
-                foreach (var artifactDefinition in artifactDefinitions)
-                    Console.WriteLine($"\tArtifact: {artifactDefinition.Artifact.Value} CLR-type: {artifactDefinition.Type.TypeString} @{artifactDefinition.Generation.Value}");
-            }
-        }
-
-        static void ThrowIfDuplicateArtifacts(ArtifactsConfiguration artifactsConfiguration)
-        {
-            var idMap = new Dictionary<ArtifactId, ClrType>();
-            bool foundDuplicate = false;
-            foreach (var artifactDefinition in artifactsConfiguration.GetAllArtifactDefinitions())
-            {
-                if (idMap.ContainsKey(artifactDefinition.Artifact))
-                {
-                    foundDuplicate = true;
-                    var artifactId = artifactDefinition.Artifact;
-                    var clrType = idMap[artifactDefinition.Artifact];
-                    
-                    ConsoleLogger.LogError($"An artifact with ArtifactId: {artifactId.Value} already exists.");
-                    ConsoleLogger.LogInfo($"Artifact with Type: {clrType.TypeString} and artifact with Type {artifactDefinition.Type.TypeString} has the same ArtifactId.");
-                }
-                else 
-                    idMap.Add(artifactDefinition.Artifact, artifactDefinition.Type);
-            }
-            if (foundDuplicate) throw new DuplicateArtifact();
-        }
-
     }
 }
